Rename HomeListItemHorizontal to HomeListItemSecondary
